+++ actions/cleanup-branches/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.3.2](https://github.com/grafana/shared-workflows/compare/cleanup-branches/v0.3.1...cleanup-branches/v0.3.2) (2026-09-09)
+
+
+### 📝 Documentation
+
+* generate action and workflow input/output tables ([#2276](https://github.com/grafana/shared-workflows/issues/2276)) ([ddcc1f8](https://github.com/grafana/shared-workflows/commit/ddcc1f84239514abefa7b02481bfcc24bd88c4ca))
+
 ## [0.3.1](https://github.com/grafana/shared-workflows/compare/cleanup-branches/v0.3.0...cleanup-branches/v0.3.1) (2026-06-10)