+++ actions/validate-policy-bot-config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.3](https://github.com/grafana/shared-workflows/compare/validate-policy-bot-config/v1.1.2...validate-policy-bot-config/v1.1.3) (2026-09-09)
+
+
+### 📝 Documentation
+
+* generate action and workflow input/output tables ([#2276](https://github.com/grafana/shared-workflows/issues/2276)) ([ddcc1f8](https://github.com/grafana/shared-workflows/commit/ddcc1f84239514abefa7b02481bfcc24bd88c4ca))
+
 ## [1.1.2](https://github.com/grafana/shared-workflows/compare/validate-policy-bot-config/v1.1.1...validate-policy-bot-config/v1.1.2) (2026-06-10)