+++ actions/validate-zizmor-config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.2.2](https://github.com/grafana/shared-workflows/compare/validate-zizmor-config/v0.2.1...validate-zizmor-config/v0.2.2) (2026-09-09)
+
+
+### 📝 Documentation
+
+* generate action and workflow input/output tables ([#2276](https://github.com/grafana/shared-workflows/issues/2276)) ([ddcc1f8](https://github.com/grafana/shared-workflows/commit/ddcc1f84239514abefa7b02481bfcc24bd88c4ca))
+
 ## [0.2.1](https://github.com/grafana/shared-workflows/compare/validate-zizmor-config/v0.2.0...validate-zizmor-config/v0.2.1) (2026-06-10)